+++ src/main/java/com/ooa1769/bs/book/support/search/naver/NaverApiProperties.java
@@ -20,6 +20,9 @@ public class NaverApiProperties {
     @Value("${naver.api.url}")
     private String url;
 
+    @Value("${naver.api.detailUrl}")
+    private String detailUrl;
+
     @Value("${naver.api.clientId}")
     private String clientId;
 
@@ -34,14 +37,29 @@ public class NaverApiProperties {
         return KAKAO_AUTHORIZATION_HEADER_CLIENT_SECRET;
     }
 
+    public String baseUrl(BookSearchParam bookSearchParam) {
+        String target = bookSearchParam.getTarget();
+        return isAllSearchUrl(target) ? url : detailUrl;
+    }
+
     public Map<String, String> queryParam(BookSearchParam bookSearchParam) {
         Map<String, String> params = new HashMap<>();
 
         int startPage = (bookSearchParam.getPage() - 1) * bookSearchParam.getSize() + 1;
-        params.put("start",  startPage+ "");
+        params.put("start",  startPage + "");
         params.put("display", bookSearchParam.getSize() + "");
         params.put("sort", bookSearchParam.getSort());
-        params.put(bookSearchParam.getTarget(), bookSearchParam.getQuery());
+
+        if (isAllSearchUrl(bookSearchParam.getTarget())) {
+            params.put("query", bookSearchParam.getQuery());
+        } else {
+            params.put(bookSearchParam.getTarget(), bookSearchParam.getQuery());
+        }
+
         return params;
+    }
+
+    private boolean isAllSearchUrl(String target) {
+        return target.equals("all");
     }
 }
