+++ src/main/java/com/ooa1769/bs/web/dto/PrevBookSearchParam.java
@@ -7,6 +7,8 @@ import lombok.Setter;
 @Setter
 public class PrevBookSearchParam {
 
+    private int prevPage;
+    private int prevSize;
     private String prevTarget;
     private String prevQuery;
     private String prevCategory;
