+++ src/main/resources/naver.properties
@@ -1,3 +1,4 @@
 naver.api.clientId=99uFekoH2IOnqitrXblg
 naver.api.clientSecret=sAtwPDzE_0
-naver.api.url=https://openapi.naver.com/v1/search/book_adv.json
+naver.api.url=https://openapi.naver.com/v1/search/book.json
+naver.api.detailUrl=https://openapi.naver.com/v1/search/book_adv.json
